user interface modification

+++ src/main/java/com/kr1sel/dto/AppUserDTO.java
@@ -5,5 +5,8 @@ import com.kr1sel.utils.Interest;
 
 import java.util.Set;
 
-public record AppUserDTO(String name, int age, String location, Set<Interest> interests) {
+public record AppUserDTO(String name,
+                         short age,
+                         String location,
+                         Set<Interest> interests) {
 }
